requirements.txt
redis==5.2.1
PyYAML==6.0.2
jinja2==3.1.4
pillow==11.0.0
requests==2.32.3
numpy==2.1.3